Move permissions and lib functions to individual files
